+++ src/main/java/com/example/api/Repository/ArticleRepository.java
@@ -3,6 +3,6 @@ package com.example.api.Repository;
 import com.example.api.Model.Article;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<Article, Float> {
+public interface ArticleRepository extends JpaRepository<Article, Integer> {
 }
